+++ pakyow-assets/spec/unit/asset_spec.rb
@@ -1,13 +1,12 @@
 RSpec.describe Pakyow::Assets::Asset do
   let :config do
     app_class = Class.new(Pakyow::Application) do
-      include Pakyow::Application::Config::Assets
-
       configurable :presenter do
         setting :path, ""
       end
     end
 
+    app_class.include_framework :assets
     app_class.config.assets
   end