+++ pakyow-realtime/lib/pakyow/realtime.rb
@@ -6,10 +6,20 @@ require "pakyow/presenter"
 
 require "pakyow/realtime/framework"
 
-require "pakyow/config/realtime"
-
 require "pakyow/application/actions/realtime/upgrader"
 
 module Pakyow
-  include Config::Realtime
+  configurable :realtime do
+    setting :server, true
+
+    setting :adapter, :memory
+    setting :adapter_settings, {}
+
+    defaults :production do
+      setting :adapter, :redis
+      setting :adapter_settings do
+        Pakyow.config.redis.to_h
+      end
+    end
+  end
 end